Module M.Children

The Children module tries to provide the flexibility of the Javascript API but in a type-safe manner by specifying exactly what can and cannot be passed to the Vnode generating function.

type t = [

| `Vnodes of Vnode.t list

| `String of string

| `Number of float

| `Bool of bool

]

Types of children that can be passed to M.v

val to_jv : t -> Jv.t