Module Mithril__Component.Lifecycle

Lifecycle Method

val on_init : (Mithril.Vnode.t -> unit) -> Mithril.Attr.attr

on_init f invokes f before the virtual DOM engine touches the vnode. See Mithril docs.

val on_create : (Mithril.Vnode.t -> unit) -> Mithril.Attr.attr

on_create f invokes f when the vnode is attached to the document. See Mithril docs.

val on_update : (Mithril.Vnode.t -> unit) -> Mithril.Attr.attr

on_update f invokes f when the vnode is updated. See Mithril docs.

val on_before_remove : (Mithril.Vnode.t -> unit) -> Mithril.Attr.attr

on_before_remove f invokes f just before the node is removed from the document. See Mithril docs.

val on_remove : (Mithril.Vnode.t -> unit) -> Mithril.Attr.attr

on_remove f invokes f when the node is removed from the document. See Mithril docs.

val on_before_update : (vnode:Mithril.Vnode.t -> old:Mithril.Vnode.t -> unit) -> Mithril.Attr.attr

on_before_update f invokes f before a vnode is diffed in an update. See Mithril docs.