Module Mithril__M.Jsonp

type opts

val opts : ?⁠params:Jv.t -> ?⁠type_:Jv.t -> ?⁠callback_name:string -> ?⁠callback_key:string -> ?⁠background:bool -> url:string -> opts

See the docs for more information about what these configuration parameters allow you to do.

val v : opts -> Jv.t Fut.or_error

v opts makes a JSON-P request to the url specified in opts. More information about this functions and some information about JSON-P.