Module Mithril__Route

module Resolver : sig ... end

type r

A simple route

val mk_route : path:string -> component:Mithril.Component.t -> r

Generate a route

type t =

| Routes of r list

| Resolver of Resolver.t

val to_jv : t -> Jv.t

Convert to JS object

type route = Brr.El.t -> string -> t -> unit

val route : route

Mithril's route function -- m.route

Static members

type set_opts

val set_opts : ?⁠params:Jv.t -> ?⁠options:Jv.t -> string -> Jv.t

val set : set_opts -> unit

Redirects to a matching route, or to the default route if no matching routes can be found. See docs.